requirements.txt
Flask-Cors==3.0.6
